requirements.txt
requests
PyYAML
python-dotenv
streamlit
pandas
